backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
keras==3.3.3
matplotlib==3.8.4
mne==1.8.0
numpy==2.1.3
Pillow==11.0.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
